feat(host-react-native): extract PluginWebView + MobilePluginHost (Phase 2c)
Lifts the remaining two non-UI primitives. After this commit
app/mobile/lib/ contains only useVerifierUrl.ts (app-specific config
persistence via Expo file system) and app/mobile/components/tlsn/ holds
only the styled approval/timeout/bottom-sheet shells — exactly the
opinionated UI a consumer would want to fork.

PluginWebView (header interception + native cookie polling)
- Moved as-is into packages/host-react-native/src/components/. Only
  external deps are RN ecosystem peer deps already declared
  (react-native-webview, @react-native-cookies/cookies as an optional
  peer). Zero internal path changes needed.

MobilePluginHost (protocol adapter)
- Moved as-is into packages/host-react-native/src/. The single
  type-only import from '../modules/tlsn-native/src' is rewritten to
  the published-package name 'tlsn-native' (matches what NativeProver
  already does and the tlsn-native.d.ts ambient decl).

Mobile consumers
- PluginScreen.tsx and assets/plugins/registry.ts now import from
  '@tlsn/host-react-native' (root) for the host adapter and
  '@tlsn/host-react-native/components' for the WebView. Mobile barrel
  index.ts re-exports everything from the package for backward compat
  with any other code paths.

host-react-native package root exports the full surface (MobilePluginHost,
handler-translation helpers, type aliases for the plugin-sdk types
adapters need). The /components, /logger, /style, /util subpaths give
consumers granular tree-shaking.

Verification: host-react-native builds clean, 33/33 tests pass, mobile
passWithNoTests passes.

Remaining Phase 2 work: extend the build-tlsn-host skill with an
Expo-flavored scaffold flow.
