+++ app/mobile/components/tlsn/index.ts
@@ -8,8 +8,8 @@ export type {
   HandlerAction,
   HandlerParams,
 } from '@tlsn/host-react-native/components';
-export { PluginWebView } from './PluginWebView';
-export type { InterceptedRequestHeader } from './PluginWebView';
+export { PluginWebView } from '@tlsn/host-react-native/components';
+export type { InterceptedRequestHeader } from '@tlsn/host-react-native/components';
 export { PluginRenderer } from '@tlsn/host-react-native/components';
 export type { DomJson, DomOptions } from '@tlsn/host-react-native/components';
 export { PluginScreen } from './PluginScreen';
