+++ packages/host-react-native/src/components/index.ts
@@ -1,5 +1,7 @@
 export { PluginRenderer } from './PluginRenderer.js';
 export type { DomJson, DomOptions, PluginRendererProps } from './PluginRenderer.js';
+export { PluginWebView } from './PluginWebView.js';
+export type { InterceptedRequestHeader } from './PluginWebView.js';
 export { NativeProver } from './NativeProver.js';
 export type {
   NativeProverHandle,
